pmic_lld: GPIO Module misrac fixes rule- R.8.4, 10.3, 10.4
1.misrac fixes rule-R.10.3 - MISRA.ETYPE.ASSIGN.2012
  An expression value of essential type unsigned char is assigned
  to an object of essential type Essentially Boolean
2.misrac fixes rule-R.8.4 - MISRA.FUNC.NOPROT.DEF.2012
  Function has a definition but no prototype
3.misrac fixes rule-R.10.3 - MISRA.ETYPE.ASSIGN.2012
  An expression value of essential type unsigned int is assigned
  to an object of essential type unsigned short int
4.misrac fixes rule-R.10.4 - MISRA.ETYPE.CATEGORY.DIFFERENT.2012
  The operands of the operator == do not have the same essential type category:
  one is signed and the other is Boolean.

+++ src/pmic_gpio_priv.h
@@ -327,11 +327,11 @@ extern "C" {
  */
 typedef struct Pmic_GpioInOutCfg_s
 {
-    uint32_t regAddr;
-    uint32_t outRegAddr;
-    uint32_t inRegAddr;
-    uint8_t  inRegBitPos;
-    uint8_t  outRegBitPos;
+    uint8_t regAddr;
+    uint8_t outRegAddr;
+    uint8_t inRegAddr;
+    uint8_t inRegBitPos;
+    uint8_t outRegBitPos;
 } Pmic_GpioInOutCfg_t;
 
 /*!
